refactor to two plugins
