+++ src/lib/remark-plugins/remark-themed-image-transform/helpers.ts
@@ -13,7 +13,7 @@ export async function getImageDimensions(src: string) {
 	} catch (e) {
 		if (e.statusCode === 404) {
 			console.error(
-				'[remarkPluginThemedImageSrcAndDimensions] Image path not found, unable to calculate dimensions ' +
+				'[remarkPluginThemedImageTransform] Image path not found, unable to calculate dimensions ' +
 					e
 			)
 			return
