+++ src/views/tutorial-view/utils/serialize-content.ts
@@ -15,7 +15,7 @@ import { TableOfContentsHeading } from 'components/table-of-contents'
 import { splitProductFromFilename } from '.'
 import remarkPluginAdjustLinkUrls from 'lib/remark-plugins/remark-plugin-adjust-link-urls'
 import { rewriteWaypointPluginsToIntegrations } from 'lib/content-adjustments'
-import remarkPluginThemedImageSrcAndDimensions from 'lib/remark-plugins/remark-image-dimensions'
+import remarkPluginThemedImageTransform from 'lib/remark-plugins/remark-themed-image-transform'
 
 export async function serializeContent(tutorial: ClientTutorial): Promise<{
 	content: MDXRemoteSerializeResult
@@ -45,7 +45,7 @@ export async function serializeContent(tutorial: ClientTutorial): Promise<{
 				[anchorLinks, { headings }],
 				paragraphCustomAlerts,
 				rewriteStaticAssetsPlugin,
-				remarkPluginThemedImageSrcAndDimensions,
+				remarkPluginThemedImageTransform,
 				[
 					remarkPluginAdjustLinkUrls,
 					{
